Checkpoint: major upgrade to: Spring 4.1.2, Spring Security 3.2.5, Hibernate 4.3.7, EhCache 2.4.3, Quartz 2.2.1.  Things compile and I can start the server and run a build.  Much is likely misconfigured and broken still, but this is a fair start.

+++ com.zutubi.pulse.master/src/java/com/zutubi/pulse/master/cleanup/config/CleanupConfiguration.java
@@ -85,7 +85,7 @@ public class CleanupConfiguration extends AbstractCleanupConfiguration
             // See if there are too many builds of our states.  We assume here
             // we are called from within the build manager (so these two dao
             // calls are within the same transaction).
-            int total = dao.getBuildCount(project, allowedStates, allowedStatuses, false);
+            int total = (int) dao.getBuildCount(project, allowedStates, allowedStatuses, false);
             if(total > retain)
             {
                 // Clean out the difference
