+++ com.zutubi.pulse.master/src/test/com/zutubi/pulse/master/agent/AgentSynchronisationServiceTest.java
@@ -561,7 +561,7 @@ public class AgentSynchronisationServiceTest extends PulseTestCase
         {
         }
 
-        public int count()
+        public long count()
         {
             return messages.size();
         }
